+++ Sources/VCore/Extensions/Core Frameworks/CGPoint.WithReversedCoordinates.swift
@@ -1,0 +1,26 @@
+//
+//  CGPoint.WithReversedCoordinates.swift
+//  VCore
+//
+//  Created by Vakhtang Kontridze on 02.04.23.
+//
+
+import CoreGraphics
+
+// MARK: - CGPoint with Reversed Coordinates
+extension CGPoint {
+    /// Returns `CGPoint` with reversed `x` and `y`.
+    ///
+    ///     let coordinates: CGPoint = .init(x: 100, y: 200)
+    ///         .withReversedCoordinates() // (x: 200, y: 100)
+    ///
+    public func withReversedCoordinates(
+        _ condition: Bool = true
+    ) -> Self {
+        if condition {
+            return CGPoint(x: y, y: x)
+        } else {
+            return self
+        }
+    }
+}